requirements.txt
requests==2.25.1
urllib3==1.26.3
Pillow==8.1.0
imgurpython==1.1.7
python-dotenv==0.17.0